refactor: split json file (for html)
